refactor(query): harden provider read-model boundary
