+++ server/src/providers/completion.ts
@@ -1,9 +1,9 @@
 import { CompletionItemKind, type CompletionItem } from "vscode-languageserver/node";
-import type { AnalysisEntry } from "../core/indexing/document-analysis-cache";
 import type { SelectorDeclHIR } from "../core/hir/style-types";
-import { getBindingDeclById, resolveBindingAtOffset } from "../core/binder/source-binding-graph";
+import { readCompletionContext } from "../core/query/read-completion-context";
 import type { CursorParams, ProviderDeps } from "./provider-deps";
 import { wrapHandler } from "./_wrap-handler";
+export { isInsideCall } from "../core/query/read-completion-context";
 
 /**
  * Handle `textDocument/completion` inside a class-util call.
@@ -37,87 +37,13 @@ function computeCompletion(params: CursorParams, deps: ProviderDeps): Completion
   }
 
   const textBefore = getTextBefore(params.content, params.line, params.character);
-  const ctx = findCompletionContext(entry, textBefore);
+  const ctx = readCompletionContext(entry, textBefore);
   if (!ctx) return null;
 
   const styleDocument = deps.styleDocumentForPath(ctx.scssModulePath);
   if (!styleDocument || styleDocument.selectors.length === 0) return null;
 
   return styleDocument.selectors.map(toCompletionItem);
-}
-
-interface CompletionContext {
-  readonly scssModulePath: string;
-}
-
-/**
- * Locate the SCSS module whose classes the cursor should complete.
- *
- * Pass 1 — cx bindings: every `classnames/bind` binding in scope
- * whose cx call is still open at the cursor.
- *
- * Pass 2 — class-util calls: for each `clsx` / `classnames` import,
- * if the cursor sits inside that call AND `textBefore` ends with
- * `<stylesVar>.<partial>` for any known style import, that style
- * import's resolved path wins.
- *
- * First hit wins; returns `null` when nothing matches.
- */
-function findCompletionContext(entry: AnalysisEntry, textBefore: string): CompletionContext | null {
-  for (const binding of entry.sourceDocument.utilityBindings) {
-    if (binding.kind !== "classnamesBind") continue;
-    const callOffset = findOpenCallOffset(textBefore, binding.localName);
-    if (callOffset === null) continue;
-    const resolution = resolveBindingAtOffset(
-      entry.sourceBindingGraph,
-      binding.localName,
-      callOffset,
-    );
-    const decl = resolution
-      ? getBindingDeclById(entry.sourceBindingGraph, resolution.declId)
-      : null;
-    if (!decl) continue;
-    if (binding.bindingDeclId !== decl.id) continue;
-    if (isInsideCall(textBefore, binding.localName)) {
-      return { scssModulePath: binding.scssModulePath };
-    }
-  }
-
-  const classUtilBindings = entry.sourceDocument.utilityBindings.filter(
-    (binding) => binding.kind === "classUtil",
-  );
-  if (classUtilBindings.length === 0 || entry.sourceDocument.styleImports.length === 0) return null;
-
-  for (const binding of classUtilBindings) {
-    const callOffset = findOpenCallOffset(textBefore, binding.localName);
-    if (callOffset === null) continue;
-    const resolution = resolveBindingAtOffset(
-      entry.sourceBindingGraph,
-      binding.localName,
-      callOffset,
-    );
-    const decl = resolution
-      ? getBindingDeclById(entry.sourceBindingGraph, resolution.declId)
-      : null;
-    if (!decl || binding.bindingDeclId !== decl.id) continue;
-    if (!isInsideCall(textBefore, binding.localName)) continue;
-    // Check if textBefore ends with `<varName>.` or `<varName>.<partial>`
-    // for any known style import. Uses simple string check instead of
-    // regex to avoid allocation in the hot completion path.
-    for (const styleImport of entry.sourceDocument.styleImports) {
-      if (styleImport.resolved.kind !== "resolved") continue;
-      const varName = styleImport.localName;
-      const dotPrefix = `${varName}.`;
-      const idx = textBefore.lastIndexOf(dotPrefix);
-      if (idx < 0) continue;
-      // Everything after `varName.` must be a partial identifier (word chars only).
-      const afterDot = textBefore.slice(idx + dotPrefix.length);
-      if (afterDot.length > 0 && !/^\w+$/.test(afterDot)) continue;
-      return { scssModulePath: styleImport.resolved.absolutePath };
-    }
-  }
-
-  return null;
 }
 
 function toCompletionItem(selector: SelectorDeclHIR): CompletionItem {
@@ -155,34 +81,6 @@ export const COMPLETION_TRIGGER_CHARACTERS = ["'", '"', "`", ",", "."] as const;
  * are ignored. Escaped quotes (backslash) are handled. This
  * means `cx(')')` correctly remains "inside" the call.
  */
-export function isInsideCall(textBefore: string, callName: string): boolean {
-  const needle = `${callName}(`;
-  const callIdx = textBefore.lastIndexOf(needle);
-  if (callIdx === -1) return false;
-
-  let depth = 1;
-  let quote: string | null = null;
-  for (let i = callIdx + needle.length; i < textBefore.length; i += 1) {
-    const ch = textBefore[i]!;
-    // Inside a quoted string — skip until the matching close quote.
-    if (quote) {
-      if (ch === quote && textBefore[i - 1] !== "\\") quote = null;
-      continue;
-    }
-    // Opening a string literal.
-    if (ch === "'" || ch === '"' || ch === "`") {
-      quote = ch;
-      continue;
-    }
-    if (ch === "(") depth += 1;
-    else if (ch === ")") {
-      depth -= 1;
-      if (depth === 0) return false;
-    }
-  }
-  return depth > 0;
-}
-
 /** All text from file start to (line, character). */
 function getTextBefore(content: string, line: number, character: number): string {
   let offset = 0;
@@ -192,12 +90,4 @@ function getTextBefore(content: string, line: number, character: number): string
     offset = nl + 1;
   }
   return content.slice(0, offset + character);
-}
-
-function findOpenCallOffset(textBefore: string, callName: string): number | null {
-  const needle = `${callName}(`;
-  const callIdx = textBefore.lastIndexOf(needle);
-  if (callIdx === -1) return null;
-  if (!isInsideCall(textBefore, callName)) return null;
-  return callIdx;
 }
